Java file
The game is meant to give the user the option to play a guessing game. The user gets 3 lives to guess which number the randomizer have chosen. 
The integer k is the lives counter if you wish to edit and try your luck with this one in a rougher difficulty.